chore

+++ lib/src/config/utils/helper_functions.dart
@@ -34,12 +34,12 @@ class ATHelperFuncs{
     => ScaffoldMessenger.maybeOf(context)?.hideCurrentSnackBar();
 
 
-  static int getRandomNumber(double ceiling){
+  static double getRandomNumber(double ceiling){
     final int number = Random().nextInt(ceiling.toInt()) + 1;
     if(number < 100){
-      return 100 + number;
+      return (100 + number).toDouble();
     }
-    return number;
+    return number.toDouble();
   }
 
   static void startTimer({
